Move PCI bus in OSTD into PCI component
