+++ kernel/comps/pci/Cargo.toml
@@ -10,6 +10,8 @@ ostd = { path = "../../../ostd" }
 component = { path = "../../libs/comp-sys/component" }
 log = "0.4"
 spin = "0.9.4"
+cfg-if = "1.0"
+bitflags = "1.3"
 
 [lints]
 workspace = true
