+++ kernel/comps/pci/src/device_info.rs
@@ -1,0 +1,51 @@
+// SPDX-License-Identifier: MPL-2.0
+
+//! PCI device Information
+use ostd::bus::pci::PciDeviceLocation;
+
+use super::cfg_space::PciDeviceCommonCfgOffset;
+
+/// PCI device ID
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct PciDeviceId {
+    /// Vendor ID
+    pub vendor_id: u16,
+    /// Device ID
+    pub device_id: u16,
+    /// Revision ID
+    pub revision_id: u8,
+    /// Programming Interface Byte
+    pub prog_if: u8,
+    /// Specifies the specific function the device performs.
+    pub subclass: u8,
+    /// Specifies the type of function the device performs.
+    pub class: u8,
+    /// Subsystem Vendor ID
+    pub subsystem_vendor_id: u16,
+    /// Subsystem ID
+    pub subsystem_id: u16,
+}
+
+impl PciDeviceId {
+    pub(super) fn new(location: PciDeviceLocation) -> Self {
+        let vendor_id = location.read16(PciDeviceCommonCfgOffset::VendorId as u16);
+        let device_id = location.read16(PciDeviceCommonCfgOffset::DeviceId as u16);
+        let revision_id = location.read8(PciDeviceCommonCfgOffset::RevisionId as u16);
+        let prog_if = location.read8(PciDeviceCommonCfgOffset::ClassCode as u16);
+        let subclass = location.read8(PciDeviceCommonCfgOffset::ClassCode as u16 + 1);
+        let class = location.read8(PciDeviceCommonCfgOffset::ClassCode as u16 + 2);
+        let subsystem_vendor_id =
+            location.read16(PciDeviceCommonCfgOffset::SubsystemVendorId as u16);
+        let subsystem_id = location.read16(PciDeviceCommonCfgOffset::SubsystemId as u16);
+        Self {
+            vendor_id,
+            device_id,
+            revision_id,
+            prog_if,
+            subclass,
+            class,
+            subsystem_vendor_id,
+            subsystem_id,
+        }
+    }
+}
