+++ kernel/comps/virtio/src/transport/pci/device.rs
@@ -3,16 +3,14 @@
 use alloc::{boxed::Box, sync::Arc};
 use core::fmt::Debug;
 
+use aster_pci::{
+    bus::PciDevice, capability::CapabilityData, cfg_space::Bar, common_device::PciCommonDevice,
+    PciDeviceId,
+};
 use aster_util::{field_ptr, safe_ptr::SafePtr};
 use log::{info, warn};
 use ostd::{
-    bus::{
-        pci::{
-            bus::PciDevice, capability::CapabilityData, cfg_space::Bar,
-            common_device::PciCommonDevice, PciDeviceId,
-        },
-        BusProbeError,
-    },
+    bus::BusProbeError,
     io::IoMem,
     irq::IrqCallbackFunction,
     mm::{DmaCoherent, HasDaddr},
