+++ kernel/comps/virtio/src/transport/pci/msix.rs
@@ -2,7 +2,8 @@
 
 use alloc::vec::Vec;
 
-use ostd::{bus::pci::capability::msix::CapabilityMsixData, irq::IrqLine};
+use aster_pci::capability::msix::CapabilityMsixData;
+use ostd::irq::IrqLine;
 
 pub struct VirtioMsixManager {
     config_msix_vector: u16,
