+++ ostd/src/bus/pci.rs
@@ -1,54 +1,12 @@
 // SPDX-License-Identifier: MPL-2.0
 
-//! PCI device Information
+//! Helper functions or structures for PCI devices.
 
 use core::iter;
 
-use super::cfg_space::PciDeviceCommonCfgOffset;
-
-/// PCI device ID
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct PciDeviceId {
-    /// Vendor ID
-    pub vendor_id: u16,
-    /// Device ID
-    pub device_id: u16,
-    /// Revision ID
-    pub revision_id: u8,
-    /// Programming Interface Byte
-    pub prog_if: u8,
-    /// Specifies the specific function the device performs.
-    pub subclass: u8,
-    /// Specifies the type of function the device performs.
-    pub class: u8,
-    /// Subsystem Vendor ID
-    pub subsystem_vendor_id: u16,
-    /// Subsystem ID
-    pub subsystem_id: u16,
-}
-
-impl PciDeviceId {
-    pub(super) fn new(location: PciDeviceLocation) -> Self {
-        let vendor_id = location.read16(PciDeviceCommonCfgOffset::VendorId as u16);
-        let device_id = location.read16(PciDeviceCommonCfgOffset::DeviceId as u16);
-        let revision_id = location.read8(PciDeviceCommonCfgOffset::RevisionId as u16);
-        let prog_if = location.read8(PciDeviceCommonCfgOffset::ClassCode as u16);
-        let subclass = location.read8(PciDeviceCommonCfgOffset::ClassCode as u16 + 1);
-        let class = location.read8(PciDeviceCommonCfgOffset::ClassCode as u16 + 2);
-        let subsystem_vendor_id =
-            location.read16(PciDeviceCommonCfgOffset::SubsystemVendorId as u16);
-        let subsystem_id = location.read16(PciDeviceCommonCfgOffset::SubsystemId as u16);
-        Self {
-            vendor_id,
-            device_id,
-            revision_id,
-            prog_if,
-            subclass,
-            class,
-            subsystem_vendor_id,
-            subsystem_id,
-        }
-    }
+/// Checks if the system has a PCI bus.
+pub fn has_pci_bus() -> bool {
+    crate::arch::pci::has_pci_bus()
 }
 
 /// PCI device Location
@@ -111,19 +69,22 @@ impl PciDeviceLocation {
 }
 
 impl PciDeviceLocation {
-    pub(super) const BIT32_ALIGN_MASK: u16 = 0xFFFC;
+    const BIT32_ALIGN_MASK: u16 = 0xFFFC;
 
-    pub(super) fn read8(&self, offset: u16) -> u8 {
+    /// Reads a 8-bit value from the PCI configuration space at the specified offset.
+    pub fn read8(&self, offset: u16) -> u8 {
         let val = self.read32(offset & Self::BIT32_ALIGN_MASK);
         ((val >> ((offset as usize & 0b11) << 3)) & 0xFF) as u8
     }
 
-    pub(super) fn read16(&self, offset: u16) -> u16 {
+    /// Reads a 16-bit value from the PCI configuration space at the specified offset.
+    pub fn read16(&self, offset: u16) -> u16 {
         let val = self.read32(offset & Self::BIT32_ALIGN_MASK);
         ((val >> ((offset as usize & 0b10) << 3)) & 0xFFFF) as u16
     }
 
-    pub(super) fn read32(&self, offset: u16) -> u32 {
+    /// Reads a 32-bit value from the PCI configuration space at the specified offset.
+    pub fn read32(&self, offset: u16) -> u32 {
         debug_assert!(
             (offset & 0b11) == 0,
             "misaligned PCI configuration dword u32 read"
@@ -131,7 +92,8 @@ impl PciDeviceLocation {
         crate::arch::pci::read32(self, offset as u32).unwrap()
     }
 
-    pub(super) fn write8(&self, offset: u16, val: u8) {
+    /// Writes an 8-bit value to the PCI configuration space at the specified offset.
+    pub fn write8(&self, offset: u16, val: u8) {
         let old = self.read32(offset & Self::BIT32_ALIGN_MASK);
         let dest = (offset as usize & 0b11) << 3;
         let mask = (0xFF << dest) as u32;
@@ -141,7 +103,8 @@ impl PciDeviceLocation {
         );
     }
 
-    pub(super) fn write16(&self, offset: u16, val: u16) {
+    /// Writes an 16-bit value to the PCI configuration space at the specified offset.
+    pub fn write16(&self, offset: u16, val: u16) {
         let old = self.read32(offset & Self::BIT32_ALIGN_MASK);
         let dest = (offset as usize & 0b10) << 3;
         let mask = (0xFFFF << dest) as u32;
@@ -151,7 +114,8 @@ impl PciDeviceLocation {
         );
     }
 
-    pub(super) fn write32(&self, offset: u16, val: u32) {
+    /// Writes an 32-bit value to the PCI configuration space at the specified offset.
+    pub fn write32(&self, offset: u16, val: u32) {
         debug_assert!(
             (offset & 0b11) == 0,
             "misaligned PCI configuration dword u32 write"
